Moving function from shard_rsm to helper files
